==> 03_preprocess.ipynb <==
## Truth Social data

import os
import pandas as pd

data_dir = "data"

csv_files = [file for file in os.listdir(data_dir) if file.endswith('.csv')]

print("CSV 파일 리스트:", csv_files)

merged_df = pd.DataFrame()

for i, csv_file in enumerate(csv_files):
    file_path = os.path.join(data_dir, csv_file)
    df = pd.read_csv(file_path)
    print(f"{i+1}/{len(csv_files)}: {csv_file} 파일 읽기 완료, 데이터 크기: {df.shape}")
    merged_df = pd.concat([merged_df, df], ignore_index=True)
    print(f"현재 병합된 데이터 크기: {merged_df.shape}")

print("최종 병합된 DataFrame 크기:", merged_df.shape)

print(merged_df)

merged_df['post_time'] = pd.to_datetime(merged_df['post_time'], errors='coerce')  # post_time을 datetime으로 변환
merged_df = merged_df.sort_values(by='post_time').reset_index(drop=True)  # post_time 기준 정렬
merged_df

merged_df = merged_df.dropna()
merged_df

merged_df.to_csv("data.csv", index=False)

import os
import pandas as pd

df = pd.read_csv('data.csv')
df

## Twitter data

UTC -> EST

from datetime import datetime
import pytz

# UTC -> EST 변환
df['date'] = pd.to_datetime(df['date'])  # 문자열을 datetime 형식으로 변환
df['date'] = df['date'].dt.tz_localize('UTC').dt.tz_convert('US/Eastern')  # UTC를 EST로 변환

# 결과 출력
df

# 시간대 정보 제거
df['date'] = df['date'].str.replace(r'-\d{2}:\d{2}$', '', regex=True)

print(df)

## X data

KST -> EST

import pandas as pd

tweet = pd.read_csv('12-26.csv')
tweet

# '·' 및 공백 제거
tweet['date'] = tweet['date'].str.replace('·', '').str.strip()

# date 열을 datetime으로 변환
tweet['date'] = pd.to_datetime(tweet['date'], format='%I:%M %p %b %d, %Y')

# 결과 출력
print(tweet)

# 시간 순 정렬
tweet = tweet.sort_values(by='date')
tweet

from pytz import timezone

# date 열을 datetime 형식으로 변환
tweet['date'] = pd.to_datetime(tweet['date'])

# KST -> EST 변환
kst = timezone('Asia/Seoul')  # KST 타임존
est = timezone('US/Eastern')  # EST 타임존

# KST 시간대를 datetime에 지정한 뒤, EST로 변환
tweet['date'] = tweet['date'].dt.tz_localize(kst).dt.tz_convert(est)

# 결과 출력
print(tweet)

# 시간대 정보를 제거
tweet['date'] = tweet['date'].dt.tz_localize(None)

# 결과 출력
print(tweet)

# df 저장
df.to_csv('est.csv', index=False)

### 3개의 데이터를 하나로 합친 후, 전처리

# 데이터프레임 합치기
combined_df = pd.concat([df, tweet], ignore_index=True)

# 결과 출력
print(combined_df)

import re

url 제거

df['text'] = df['text'].apply(lambda x: re.sub(r'http\S+', '', x))
df

# 'text' 열에서 공백 또는 빈 문자열 제거
df = df[df['text'].str.strip() != '']

# 인덱스 재설정
df = df.reset_index(drop=True)
df
