==> 06_data join.ipynb <==
Trump Tweet Dataset

import pandas as pd
df = pd.read_csv("C:/Users/seonahryu/Desktop/urp/tweet_sentiment_prediction.csv")

Nasdaq Dataset

import pandas as pd
df_nasdaq = pd.read_csv("C:/Users/seonahryu/Desktop/urp/nasdaq_data_2009_2024.csv", index_col = 'Unnamed: 0')

import seaborn as sns
import matplotlib.pyplot as plt

plt.figure(figsize=(16,9))
sns.lineplot(y=df_nasdaq['Close'], x=df_nasdaq.index)
plt.xlabel('time')
plt.ylabel('close price')

df index를 prediction_time으로 설정

df.set_index('prediction_date', inplace=True)
df.index = pd.to_datetime(df.index)
df.head(5)

df = df.drop(columns=["date", "text"])
df

df_nasdaq에서 전날 종가로 다음날 종가 예측

# Close 열만 선택하여 새로운 데이터프레임 저장
df_close = df_nasdaq[['Close']].copy()
df_close

import pandas as pd

# 인덱스를 datetime 형식으로 변환
df_close.index = pd.to_datetime(df_close.index)

# 1일씩 다음날로 이동
df_close.index = df_close.index + pd.DateOffset(days=1)

# 결과 확인
df_close

inner join (prediction_date가 index인 df와 df_close)

# df의 인덱스와 df_close의 인덱스가 일치하는 행만
in_df = df.join(df_close, how='inner')

in_df.head(10)

in_df.rename(columns={'Close': 'past_day_close'}, inplace=True)

nasdaq index(date) = in_df index(prediction_date) -> inner join

df_nasdaq.index = pd.to_datetime(df_nasdaq.index)

df_nasdaq = df_nasdaq.drop(columns=["Close"]) #Adj Close 예측
df

# NaN이 없도록 inner join
df_combined = in_df.join(df_nasdaq, how='inner')

print(df_combined.head(10))

# 인덱스를 열로 변환하고 열 이름 지정
df_combined.reset_index(drop=False, inplace=True)  # drop=False로 인덱스를 열로 변환
df_combined.rename(columns={'index': 'date'}, inplace=True)  # 열 이름 변경

# CSV 파일로 저장
df_combined.to_csv('combined_data.csv', index=False)  # 인덱스는 저장하지 않음
